requirements.txt
attrs==22.1.0
aws-cdk-lib==2.15.0
boto3==1.26.2
botocore==1.29.2
cattrs==22.2.0
constructs==10.1.149
exceptiongroup==1.0.1
jmespath==1.0.1
jsii==1.70.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.12
